Refactor geeda.apply() to return printouts
