+++ src/geeda/column/is_categorical.py
@@ -1,4 +1,3 @@
-from calendar import c
 import pandas as pd
