--- src/geeda/dataframe/check_inf.py
@@ -1,16 +1,0 @@
-import pandas as pd
-
-
-def check_inf(column: pd.Series) -> int:
-    """
-    Quantify the amount of infinite values present in the given column.
-
-    Args:
-        column (pd.Series):
-            The column to analyze
-
-    Returns:
-        int:
-            The number of infinite values present
-    """
-    pass
