--- src/geeda/dataframe/check_nan.py
@@ -1,2 +1,0 @@
-def check_nan():
-    pass
